refactor(recovery): split structure capture and apply modules
